changes and more examples

+++ mypy/LINTING/example2.py
@@ -1,6 +1,6 @@
-#code made in class with teacher ariel
-#A01753176 Gilberto André García Gaytán
-x: list[int] = [1, 2, 3]
+# code made in class with teacher Ariel.
+# A01753176.
+x: list[float] = [1, 2, 3]
 x.append(5)
 x.append(3.1416)
 print(x)
